Cast __LINE__ to unsigned int and print it with %u.
No version of ISO C I can find (from C90 to C18) says anything more
about __LINE__ than that it's an "integer constant", with no indication
whether it's signed or unsigned (or whether it's int or long or long
long), so we just cast it to "unsigned int" and print it with %u, as it
would make Absolutely No Sense if it were negative.

This should squelch some invalidPrintfArgType_uint warnings from
cppcheck.

+++ epan/dfilter/semcheck.c
@@ -33,7 +33,7 @@
 
 #ifdef DEBUG_dfilter
 #define DebugLog(x) \
-	ws_debug_printf("%s:%u: ", __FILE__, __LINE__); \
+	ws_debug_printf("%s:%u: ", __FILE__, (unsigned int)__LINE__); \
 	ws_debug_printf x; \
 	fflush(stdout)
 #else
